+++ src/HetrickCV.cpp
@@ -40,6 +40,7 @@ void init(rack::Plugin *p) {
 	p->addModel(modelMinMax);
 	p->addModel(modelPhaseDrivenSequencer);
 	p->addModel(modelPhaseDrivenSequencer32);
+	p->addModel(modelPhasorAnalyzer);
 	p->addModel(modelPhasorDivMult);
 	p->addModel(modelPhasorEuclidean);
 	p->addModel(modelPhasorGates);